Small tweak to use the "skip" functionality for cookie banner
Previously if the cookie banner was shown in error, it would
briefly be rendered on the page before being hidden.

This tweaks the done() callback in the cookie custom setup to
return { skip: true } if we do not want to render the message.

This has the benefit of allowing us to track incorrect message
allocataions.

 🐿 v2.8.0